refactor: ContentRatingConverter에 assert 추가 및 기본값 적용
- 불필요한 NPE를 줄이기 위해 코드 개선
- ContentRating이나 DB 저장 값이 null이 아닌지 assert 검증 추가
- 불러온 DB 저장 값이 null일 경우 기본 값 적용
