+++ movie-domain/src/main/java/com/example/domain/converter/ContentRatingConverter.java
@@ -1,0 +1,25 @@
+package com.example.domain.converter;
+
+import com.example.domain.model.valueObject.ContentRating;
+import jakarta.persistence.AttributeConverter;
+import jakarta.persistence.Converter;
+
+@Converter(autoApply = true) // 모든 ContentRating 필드에 자동으로 적용
+public class ContentRatingConverter implements AttributeConverter<ContentRating, String> {
+
+    private static final ContentRating DEFAULT_RATING = ContentRating.UNKNOWN;
+
+    @Override
+    public String convertToDatabaseColumn(ContentRating contentRating) {
+        assert contentRating != null : "[ERROR] ContentRating is null.";
+        return contentRating.getDbValue();
+    }
+
+    @Override
+    public ContentRating convertToEntityAttribute(String dbData) {
+        assert dbData != null : "[ERROR] ContentRating from DB value is null.";
+        return ContentRating.fromDbValue(dbData)
+                .orElse(DEFAULT_RATING);
+    }
+
+}
